requirements.txt
flask
Flask-Cors
psycopg
psycopg-binary
python-dotenv